refactor(Tests API Routes): 正交拆分 test_api_document_routes.py（937行）为 corpus/pipelines/documents 三个域模块;
- test_api_corpus.py: Corpus CRUD 与 chunking 策略序列化（4 tests）
- test_api_pipelines.py: Pipeline API 与 OpenAPI schema 验证（5 tests）
- test_api_documents.py: 文档分块、同步、重建、摄入、搜索路由（12 tests）
- Fake 对象统一从 conftest 导入，消除 194 行重复定义
